refactor: replace bare unsigned with int throughout codebase
Replace all occurrences of bare `unsigned` (unsigned int) with `int`
across all headers and implementation files. Preserved unsigned char
(Color/Byte types), unsigned short/long (BMP binary format structs).

Added (int) casts where needed to fix -Wsign-compare warnings when
comparing int variables against container::size() return values.
Changed 1U/2U literals to 1/2 to resolve overload ambiguity after
removing unsigned overloads.

+++ code/game/controller/buy_now_controller.cc
@@ -33,7 +33,7 @@ void BuyNowController::RandomizeVelocity(fPoint &fVel, fPoint pPos) {
 }
 
 void BuyNowController::DrawSlimes() {
-  for (unsigned i = 0; i < mSlimes.size(); i++)
+  for (int i = 0; i < (int)mSlimes.size(); i++)
     mSlimes[i]->Draw(pGl->GetDrawer());
 }
 
@@ -41,12 +41,12 @@ void BuyNowController::Update() {
   EntityListController::Update();
 
   if (tVel.Tick()) {
-    for (unsigned i = 0; i < mSlimes.size(); i++)
+    for (int i = 0; i < (int)mSlimes.size(); i++)
       if (float(rand()) / RAND_MAX < .25)
         RandomizeVelocity(mSlimeVel[i], mSlimePos[i]);
   }
 
-  for (unsigned i = 0; i < mSlimes.size(); i++) {
+  for (int i = 0; i < (int)mSlimes.size(); i++) {
     mSlimes[i]->Update();
     mSlimePos[i] += mSlimeVel[i];
     mSlimes[i]->pos = mSlimePos[i].ToPnt();
