+++ code/game_utils/sound_sequence.cc
@@ -3,7 +3,7 @@
 SoundSequence::SoundSequence() : nActive(0) {}
 
 bool SoundSequence::Toggle() {
-  if (nActive == vSounds.size() - 1) {
+  if (nActive == (int)vSounds.size() - 1) {
     nActive = 0;
     return true;
   }
@@ -13,13 +13,13 @@ bool SoundSequence::Toggle() {
 
 Index SoundSequence::GetSound() { return vSounds[nActive]; }
 
-unsigned SoundSequence::GetTime() const {
+int SoundSequence::GetTime() const {
   if (vIntervals.empty())
     return 1;
   return vIntervals[nActive];
 }
 
-void SoundSequence::Add(Index iSnd, unsigned nTime) {
+void SoundSequence::Add(Index iSnd, int nTime) {
   vSounds.push_back(iSnd);
   vIntervals.push_back(nTime);
 }
